pyproject.toml
[tool.poetry]
name = "hackathons canada"
version = "0.4.5"
description = "A platform to connect Hackers to Hackathons"
authors = ["Jason <git@jasoncameron.dev>"]
license = "GPLv3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
docker = "^7.1.0"
django-crispy-forms = "^2.2"
crispy-bootstrap4 = "^2024.1"
flower = "^2.0.1"
celery = { extras = ["redis"], version = "^5.4.0" }
django-allauth = { extras = ["socialaccount"], version = "0.63.5" }
django-countries = { extras = ["pyuca"], version = "^7.6.1" }
pillow = "^10.4.0"
django-unfold = "^0.30.0"
django-celery-beat = "^2.7.0"
cloudscraper = "^1.2.71"
bs4 = "^0.0.2"
django-ratelimit = "^4.1.0"
django-redis = {extras = ["hiredis"], version = "^5.4.0"}
python-dotenv = "^1.0.1"
django-filter = "^24.3"
icalendar = "^6.1.0"
django-silk = "^5.3.2"
discord-py = "^2.4.0"
audioop-lts = { version = "^0.2.1", python = "^3.13" }
dj-database-url = "^2.3.0"
django-split-settings = "^1.3.2"
python-decouple = "^3.8"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.1"}
ruff = "^0.5.1"
djhtml = "^3.0.6"
celery-types = "^0.22.0"
pre-commit = "^3.7.1"

[tool.poetry.group.prod]
optional = true

[tool.poetry.group.prod.dependencies]
gunicorn = "^20.1.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}


[tool.mypy]
plugins = ["mypy_django_plugin.main"]
exclude = ["migrations", "node_modules", "static", "venv", ".idea"]

[[tool.mypy.overrides]]
module = "*"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "hackathons_canada.settings"


[tool.ruff]
exclude = ["migrations", "node_modules", "static", "venv", ".idea", ".git"]
fix = true
show-fixes = true

[tool.ruff.format]
docstring-code-format = true
line-ending = "lf"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
